+++ packages/auth/src/entity/AccessToken.ts
@@ -4,10 +4,18 @@ export interface AccessToken {
   client_id?: string;
   user_id?: string;
   scope?: string;
+  is_admin?: boolean;
 }
 
 export interface TokenRepo {
-  save: (option: { key: string; value: AccessToken; useDefaultExpiry: boolean }) => Promise<string>;
-  find: (option: { key: string }) => Promise<AccessToken>;
-  deleteToken: (option: { key: string }) => Promise<any>;
+  save: (option: {
+    user_id: string;
+    access_token: string;
+    useDefaultExpiry: boolean;
+    client_id?: string;
+    is_admin?: boolean;
+  }) => Promise<string>;
+  find: (token: string) => Promise<AccessToken>;
+  findByUserId: (id: string) => Promise<AccessToken[]>;
+  deleteToken: (user_id: string, token: string) => Promise<any>;
 }