+++ packages/ui-control/pages/_app.tsx
@@ -1,8 +1,8 @@
 import { ApolloProvider } from '@apollo/client';
+import { AuthProvider, AlertProvider } from 'components';
 import { NextPage } from 'next';
 import React from 'react';
-import { AuthProvider, AlertProvider } from '../components';
-import { useApollo } from '../utils/apolloClient';
+import { useApollo } from 'utils/apolloClient';
 
 const App: NextPage<any> = ({ Component, pageProps }) => {
   const apolloClient = useApollo(pageProps.initialApolloState);