+++ packages/ui-control/types/apolloContext.ts
@@ -3,5 +3,6 @@ import { Response } from 'express';
 export type ApolloContext = {
   authUri: string;
   res: Response;
-  token: string;
+  refreshToken?: string;
+  accessToken?: string;
 };