+++ packages/ui-control/types/authentication.ts
@@ -3,5 +3,5 @@ import { User } from './user';
 export type Authentication = {
   loading: boolean;
   loggedIn: boolean;
-  user: User | null | undefined;
+  user?: User | null | undefined;
 };