+++ packages/ui-control/types/index.ts
@@ -5,3 +5,6 @@ export * from './authentication';
 export * from './registerResponse';
 export * from './apolloContext';
 export * from './loginResponse';
+export * from './inMemoryToken';
+export * from './refreshTokenResponse';
+export * from './updateProfileResponse';