+++ packages/ui-control/types/updateProfileResponse.ts
@@ -1,0 +1,5 @@
+export interface UpdateProfileResponse {
+  ok: boolean;
+  username: string;
+  email: string;
+}