+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 - parametric search
 - apollo-subscription
 - redis-based queryDatabase
+- add refreshToken in auth server and ui-control
+- add synchroize logout
 
 ### Modified
 - dev-net to using queryHandler